Move PINKeypad to component root
