+++ packages/client/src/views/Unlock/Unlock.tsx
@@ -21,7 +21,7 @@ import { IUserDetails, getUserName } from '@client/utils/userUtils'
 import { pinValidator } from '@client/views/Unlock/ComparePINs'
 import { ErrorMessage } from '@opencrvs/components/lib/ErrorMessage'
 import { Logout } from '@opencrvs/components/lib/icons'
-import { PINKeypad } from '@opencrvs/components/lib/interface'
+import { PINKeypad } from '@opencrvs/components/lib/PINKeypad'
 import { GQLHumanName } from '@opencrvs/gateway/src/graphql/schema'
 import * as React from 'react'
 import * as ReactDOM from 'react-dom'
