+++ packages/components/src/index.ts
@@ -29,6 +29,7 @@ export * from './Link'
 export * from './List'
 export * from './Loader'
 export * from './PDFViewer'
+export * from './PINKeypad'
 export * from './PageWrapper'
 export * from './Pagination'
 export * from './PasswordInput'
