+++ packages/components/src/interface/index.ts
@@ -20,7 +20,6 @@ export * from './ListViewSimplified/ListViewSimplified'
 export * from './DataTable/SortAndFilter'
 export * from './SectionDrawer'
 export * from './Divider'
-export * from './PINKeypad'
 export * from './SearchTool'
 export * from './ExpandableNotification/ExpandableNotification'
 export * from './Header/AppHeader'
